Chore: point the examples at the moved templates
