+++ example/tanstack-start/src/routes/index.tsx
@@ -1,7 +1,7 @@
 import ImageResponse from "takumi-js/response";
 import { createFileRoute } from "@tanstack/react-router";
 import { Axe } from "lucide-react";
-import DocsTemplate from "../../../../takumi-template/src/templates/docs-template";
+import DocsTemplate from "../../../../docs/app/registry/image/docs";
 
 export const Route = createFileRoute("/")({
   server: {
